enhancements to rule parsing for reuse of custom builtins
this commit adds a "with builtins" clause to the end of an "@import" directive
within a Rule so that the rule parser knows whether to load the rule with its
own set of builtins or to use BuiltinRegistry.theRegistry (default). The latter
is enforced for predefined rules (e.g. RDFS, OWL). Sets of rule files can now
share a common BuiltinRegistry instance.